Frame touch functionality & bug fix
I added frame touch functionality so that when you click on a frame at the top of the page, it will register what is needed for editing shots. This is the first part for editing shots. I also fixed a bug I found where on reload, the old pinstates wouldnt load properly.